+++ zest/releaser/longtest.py
@@ -3,44 +3,42 @@
 from __future__ import unicode_literals
 
 import logging
-import os
 import sys
 import tempfile
 import webbrowser
 
-from six.moves import shlex_quote
+try:
+    from readme.rst import render
+    HAVE_README = True
+except ImportError:
+    HAVE_README = False
 
 from zest.releaser import utils
-from zest.releaser.utils import execute_command
+from zest.releaser.utils import _execute_command
 
 logger = logging.getLogger(__name__)
 
 
 def show_longdesc():
-    filename1 = tempfile.mktemp()
-    filename2 = tempfile.mktemp()
-    filename2 = filename2 + '.html'
-    # Note: for the setup.py call we use execute_command() from our
-    # utils module. This makes sure the python path is set up right.
-    # For the other calls we use os.system(), because that returns an
-    # error code which we need.
-    execute_command(['sh', '-c', utils.cmd_to_text(
-        utils.setup_py('--long-description > %s' % filename1))])
-    error = os.system('rst2html.py %s > %s' % (shlex_quote(filename1),
-                                               shlex_quote(filename2)))
-    if error:
-        # On Linux it needs to be 'rst2html', without the '.py'
-        error = os.system('rst2html %s > %s' % (shlex_quote(filename1),
-                                                shlex_quote(filename2)))
-    if error:
-        # Alternatively, zc.rst2 provides rst2 xyz.
-        error = os.system('rst2 html %s > %s' % (shlex_quote(filename1),
-                                                 shlex_quote(filename2)))
-    if error:
+    if not HAVE_README:
         logging.error(
-            'Error generating html. Please install docutils (or zc.rst2).')
+            'Error generating html. Please install `readme`.')
         sys.exit(1)
-    url = 'file://' + filename2
+
+    filename = tempfile.mktemp('html')
+    # Note: for the setup.py call we use _execute_command() from our
+    # utils module. This makes sure the python path is set up right.
+    longdesc = _execute_command(utils.setup_py('--long-description'))
+    html, rendered = render(longdesc)
+    if not rendered:
+        logging.error(
+            'Error generating html. Invalid ReST.')
+        sys.exit(1)
+
+    with open(filename, 'wb') as fh:
+        fh.write(html.encode('utf-8'))
+
+    url = 'file://' + filename
     logging.info("Opening %s in your webbrowser.", url)
     webbrowser.open(url)